requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
redis[hiredis]==5.0.3
python-dotenv==1.0.1